iio: buffer-dma: Add helper functions
This patch adds some helper functions to access private member of the
buffer dma engine. This is an introductory patch with the goal of split
buffer-dma.h into a public and private header.

+++ include/linux/iio/buffer_impl.h
@@ -30,7 +30,7 @@ struct iio_buffer_block {
 	__u32 size;
 	__u32 bytes_used;
 	__u32 type;
-	__u32 flags;
+	unsigned long flags;
 	union {
 		__u32 offset;
 	} data;
